Refactor variant classification functions into variants.ts
